handle FFE changes to interface map
FFE automatically generates new logical device and interface map
graph nodes to accomodate interface speed changes like plugging in
a 1G server to an interface initialy provisioned for 10G operation.

The `apstra_device_allocation` resource would previously have
detected state drift which required remediation. Now, the new
interface_map's relationship with the original is recognized by
parsing the `label` field. When the label indicates that the
interface_map ID change is due to FFE, the old interface map ID
is copied into the state so that no config drift registers.

It is important to note that once an FFE event has transformed
a switch, it has a new Logical Device which no longer can be
associated with the original catalog item. Destroying and
re-creating the resource is no longer possible. This isn't a
terraform (or provider) issue, it's just how Apstra works. This
fact made it important to do in-place updates to the resource,
so changes to device_key no longer require resoruce replacement.